02_ChatGPT/네이버이메일자동화프로그램/requirements.txt
selenium==4.15.2
mysql-connector-python==9.3.0
webdriver-manager==4.0.1